Remove RecordParser::Sorter
Once the RecordParser::Sorting classes were left with no responsibility
other than sorting, the RecordParser::Sorter class became redundant.

This commit removes RecordParser::Sorter and splits its behavior
among the RecordParser::Sorting classes' #sort methods. Specs are also
adapted in this commit, so that the change is green-to-green.
