+++ lib/record_parser.rb
@@ -1,4 +1,3 @@
 require 'record_parser/application'
 require 'record_parser/record'
-require 'record_parser/sorter'
 require 'record_parser/sorting'
