+++ lib/record_parser/sorting/base.rb
@@ -4,12 +4,6 @@ module RecordParser
       def initialize(records)
         @records = records
       end
-
-    private
-
-      def sorter
-        Sorter.new(@records)
-      end
     end
   end
 end
