+++ lib/record_parser/sorting/by_birth_date.rb
@@ -2,7 +2,7 @@ module RecordParser
   module Sorting
     class ByBirthDate < Base
       def sort
-        sorter.by_birth_date
+        @records.sort_by(&:birth_date)
       end
     end
   end
