+++ lib/record_parser/sorting/by_gender_and_last_name.rb
@@ -2,7 +2,13 @@ module RecordParser
   module Sorting
     class ByGenderAndLastName < Base
       def sort
-        sorter.by_gender_and_last_name
+        gender_sets.flat_map { |same_gender| same_gender.sort_by(&:last_name) }
+      end
+
+    private
+
+      def gender_sets
+        @records.partition(&:feminine?)
       end
     end
   end
