+++ spec/record_parser/sorting/by_birth_date_spec.rb
@@ -1,12 +1,23 @@
 require 'spec_helper'
 require 'support/shared_examples_for_sortings'
+require 'support/stub_helper'
 
 module RecordParser
   module Sorting
     describe ByBirthDate do
+      let(:sorting) { ByBirthDate.new(records) }
       describe '#sort' do
-        it_has_behavior 'sorts records', :by_birth_date do
-          let(:sorting_class) { ByBirthDate }
+        subject { sorting.sort }
+        context 'with only one record' do
+          let(:messages) { %i(birth_date) }
+          it_has_behavior 'returns the single record'
+        end
+        context 'with two records' do
+          before do
+            prepare(first , { birth_date: date_from('9/14/1953') })
+            prepare(second, { birth_date: date_from('9/14/1984') })
+          end
+          it_has_behavior 'returns the two in correct order'
         end
       end
     end
